Question List
When did the participate log in?
Which month did the participant log in?
Which week did the participant log in?
Which day did the participants log in?
Which hour did the participants log in?
What is the duration of their log in on different time (month, day, hour)?
What is the gender of the participants log in on different time (month, day, hour)?
What is the age of the participants log in on different time (month, day, hour)?
Are the people log in on different time (month, day, hour) primary participants?